LangGraph/LangGraph_Human_in_the_loop/pyproject.toml
[project]
name = "langgraph-human-in-the-loop-financial"
version = "0.1.0"
description = "LangGraph Human-in-the-loop agent for financial document analysis with MCP tools"
authors = [
    {name = "Aleksei Kolesnikov", email = "kolesnikov.aliaksei@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langgraph>=0.2.0",
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.0.1",
    "mcp>=1.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "arize-phoenix-otel>=0.6.0",
    "arize-phoenix-client>=1.29.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "opentelemetry-sdk>=1.20.0",
    "pypdf>=4.0.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]

[tool.uv]
package = false

[tool.uv.workspace]
members = ["mcp_server"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]
ignore = ["D100", "D104"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
